% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcl2fastq.R
\name{bcl2fastq}
\alias{bcl2fastq}
\title{Generating fastq running bcl2fastq}
\usage{
bcl2fastq(samplesheet, output.folder, threadBCL = 1,
  threadWritingFastq = 1, threadDem = 6)
}
\arguments{
\item{samplesheet, }{a character string indicating the Illumina folder where the Samplesheet.csv is located.}

\item{output.folder, }{a character string indicating the fastq output folder.}

\item{threadBCL, }{a number indicating the number of threads to be used for loading BCL data}

\item{threadWritingFastq, }{a number indicating the number of threads to be used for writing FASTQ data
@param threadDem, a number indicating the number of threads to be used for for processing demultiplexed data}
}
\value{
Fastq files
}
\description{
This function executes the Illumina bcl2fastq program
}
\examples{
\dontrun{
    #running rsemstar index for human
    demultiplexing("~/home/data/input","~/home/data/output")

}
}
\author{
Marco Beccuti, marco.beccuti [at] unito [dot] it, University of Torino
}
