% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerRun.R
\name{dockerRun}
\alias{dockerRun}
\title{Run docker container}
\usage{
dockerRun(params = NULL)
}
\arguments{
\item{params, }{a character string containing all parameters needed to run the docker container.}
}
\value{
0 if success, 1 if parameters are missing, 2 if dockerid file is present, 3 if docker execution fails.
}
\description{
This is an internal function executing a docker container. Not to be used by users.
}
\examples{
\dontrun{
    #running runDocker
     dockerRun(params=NULL)

}
}
\author{
Marco Beccuti, marco.beccuti [at] unito [dot] it, University of Torino
}
